Added new Player classes, but haven't implemented them yet.
